Split Path into an expect/actual class
I want to move methods into Path to make use from Java nicer.
